plaid-manager/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
Django==3.2.6
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-six==1.0.4
djangorestframework==3.12.4
idna==3.2
kombu==5.1.0
plaid-python==7.5.0
prompt-toolkit==3.0.19
pycparser==2.20
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
xlwt==1.3.0
